Use "Pump" to transfer http bodies when forwarding http-requests (and their responses) to backends (or to some other Gateleen verticles). This solves OutOfMemoryErrors huge http-bodies in case of fast incomming and slow draining
